+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  
  [Changed]
  - Changed code style to PSR-2
+ - addInstitution URL nad email in disco-tpl.php are loaded from a config file
  
  ## [v1.4.0]
  [Added]